Refactor: History Directory
