+++ green/src/main/java/com/greengrim/green/core/history/service/GetHistoryService.java
@@ -1,10 +1,13 @@
-package com.greengrim.green.core.history;
+package com.greengrim.green.core.history.service;
 
 
 import static com.greengrim.green.common.entity.Time.calculateTime;
 
 import com.greengrim.green.common.entity.dto.PageResponseDto;
+import com.greengrim.green.core.history.entity.History;
+import com.greengrim.green.core.history.entity.HistoryOption;
 import com.greengrim.green.core.history.dto.HistoryResponseDto.HistoryInfo;
+import com.greengrim.green.core.history.repository.HistoryRepository;
 import jakarta.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
@@ -16,19 +19,9 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class HistoryService {
+public class GetHistoryService {
 
     private final HistoryRepository historyRepository;
-
-    @Transactional
-    public void register(Long memberId, Long targetId, String title, String imgUrl,
-                     HistoryOption historyOption, int point, int totalPoint) {
-        History history = History.builder()
-                .memberId(memberId).targetId(targetId).title(title).imgUrl(imgUrl)
-                .historyOption(historyOption).point(point).totalPoint(totalPoint)
-                .build();
-        historyRepository.save(history);
-    }
 
     public PageResponseDto<List<HistoryInfo>> getMyHistory(Long id, int page, int size) {
         Page<History> histories = historyRepository.findByMemberId(id, PageRequest.of(page, size, Direction.DESC, "createdAt"));
